update packages and profiles
