+++ OLGA/OLGA-Core/src/test/java/semanticstore/ontology/library/generator/generators/csharp/OLGAOutputOptionTestCsharp.java
@@ -24,46 +24,30 @@
  */
 package semanticstore.ontology.library.generator.code.generator.generators.csharp;
 
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.io.UnsupportedEncodingException;
+import static org.junit.Assert.assertTrue;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
+import org.junit.BeforeClass;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import semanticstore.ontology.library.generator.exceptions.InvalidUriException;
 import semanticstore.ontology.library.generator.global.CODE;
 import semanticstore.ontology.library.generator.global.LIBRARY;
 import semanticstore.ontology.library.generator.service.OlgaService;
+import semanticstore.ontology.library.generator.test.utils.GeneratedOntologies;
 
-public class WrongNameTest {
+public class OLGAOutputOptionTestCsharp {
 
-  @Rule
-  public ExpectedException expected = ExpectedException.none();
 
-  private final ByteArrayOutputStream errContent = new ByteArrayOutputStream();
+  static Path outputPath;
 
-  @Before
-  public void setUpStreams() throws UnsupportedEncodingException {
-    PrintStream printStream = new PrintStream(errContent, false, "UTF-8");
-    System.setErr(printStream);
-    errContent.reset();
-  }
+  @BeforeClass
+  public static void setUpStreams() throws Exception {
 
-  @After
-  public void cleanUpStreams() {
-    System.setErr(null);
-
-  }
-
-  @Test
-  public void testWrongName() throws Exception {
-
-    expected.expect(InvalidUriException.class);
+    outputPath = Files.createTempDirectory("testOutput");
+    outputPath.toFile().deleteOnExit();
 
     OlgaService service = new OlgaService();
 
@@ -72,15 +56,26 @@ public class WrongNameTest {
     inputCmdParameters.put("library", LIBRARY.TRINITY);
     inputCmdParameters.put("skipInverseRelations", false);
     inputCmdParameters.put("skipCompile", true);
-    inputCmdParameters.put("ontVersion", "01.99.45");
+    inputCmdParameters.put("out", outputPath.toFile().toString());
 
     String resourcesDirectory = null;
-    resourcesDirectory =
-        Paths.get(WrongNameTest.class.getClassLoader().getResource("simple/wrongname.owl").toURI())
-            .toFile().toString();
-    inputCmdParameters.put("name", "WrongNameOntology");
+    resourcesDirectory = Paths
+        .get(GeneratedOntologies.class.getClassLoader().getResource("simple/simple.owl").toURI())
+        .toFile().toString();
+    inputCmdParameters.put("name", "testSimpleWithOut");
     inputCmdParameters.put("pathToOntologiesParam", resourcesDirectory);
     service.invokeOlga(inputCmdParameters);
+  }
 
+  @After
+  public void cleanUpStreams() {
+    System.setErr(null);
+  }
+
+  @Test
+  public void testFolderExists() {
+    assertTrue(Files.exists(outputPath.resolve("testSimpleWithOut-dotnetTrinity")));
+    assertTrue(Files.exists(
+        outputPath.resolve("testSimpleWithOut-dotnetTrinity").resolve("testsimplewithout.csproj")));
   }
 }
